HIA-1635 Move JsonConfig
